requirements.txt
# Core dependencies
fastmcp>=0.4.0  # Base MCP SDK
pydantic>=2.0.0
websockets>=10.0.0
aiohttp>=3.8.0
PyYAML>=6.0.0

# Framework integrations
Django>=3.2.0
Flask>=2.0.0
fastapi>=0.70.0
uvicorn>=0.15.0

# Authentication
PyJWT>=2.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
SQLAlchemy>=1.4.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
httpx>=0.23.0

# Development
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.950

# Documentation
Sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0
myst-parser>=0.18.0

# Optional dependencies
python-multipart>=0.0.5  # For FastAPI form data
django-cors-headers>=3.10.0  # For Django CORS
flask-sqlalchemy>=3.0.0  # For Flask SQLAlchemy
flask-jwt-extended>=4.0.0  # For Flask JWT
